requirements.txt
opensearch-py==2.8.0
python-dotenv==1.0.1
schedule==1.2.2
certifi==2025.1.31